Made the clock sample happy when run in checked mode. Review URL: https://chromiumcodereview.appspot.com//9432005
git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@4406 260f80e4-7a28-3924-810f-c04153c831b5

+++ dart/samples/clock/Ball.dart
@@ -16,16 +16,16 @@ class Ball {
 
   Element root;
   HTMLImageElement elem;
-  int x, y;
+  double x, y;
   double vx, vy;
   double ax, ay;
   double age;
 
-  Ball(Element this.root, int this.x, int this.y, int color) {
+  Ball(Element this.root, double this.x, double this.y, int color) {
     elem = window.document.createElement('img');
     elem.src = Balls.PNGS[color];
     Util.abs(elem);
-    Util.pos(elem, x, y);
+    Util.pos(elem, x.toDouble(), y.toDouble());
     root.appendChild(elem);
 
     ax = 0.0;
@@ -41,8 +41,8 @@ class Ball {
     vx += ax * delta;
     vy += ay * delta;
 
-    x += vx * delta;
-    y += vy * delta;
+    x += (vx * delta).toInt();
+    y += (vy * delta).toInt();
 
     // Handle falling off the edge.
     if ((x < RADIUS) || (x > Util.clientWidth())) {
