+++ dart/samples/clock/Colon.dart
@@ -5,21 +5,21 @@
 class Colon {
   Element root;
 
-  Colon(int pos) {
+  Colon(double pos) {
     root = window.document.createElement('div');
     Util.abs(root);
-    Util.pos(root, pos, 0);
+    Util.pos(root, pos, 0.0);
 
     HTMLImageElement dot = window.document.createElement('img');
     root.appendChild(dot);
     Util.abs(dot);
-    Util.pos(dot, 0, 2 * CountDownClock.BALL_HEIGHT);
+    Util.pos(dot, 0.0, 2.0 * CountDownClock.BALL_HEIGHT);
     dot.src = Balls.PNGS[4];
 
     dot = window.document.createElement('img');
     root.appendChild(dot);
     Util.abs(dot);
-    Util.pos(dot, 0, 4 * CountDownClock.BALL_HEIGHT);
+    Util.pos(dot, 0.0, 4.0 * CountDownClock.BALL_HEIGHT);
     dot.src = Balls.PNGS[4];
   }
 }
